clean up and subdomain validation
